+++ xskillscore/tests/test_skipna_functionality.py
@@ -1,7 +1,7 @@
 import dask
 import numpy as np
 import pytest
-from xarray.tests import assert_allclose
+from xarray.tests import CountingScheduler, assert_allclose, raise_if_dask_computes
 
 from xskillscore.core.deterministic import (
     mae,
@@ -33,33 +33,6 @@ WEIGHTED_METRICS = [
 ]
 
 NON_WEIGHTED_METRICS = [median_absolute_error]
-
-
-class CountingScheduler:
-    """Simple dask scheduler counting the number of computes.
-    Reference: https://stackoverflow.com/questions/53289286/"""
-
-    def __init__(self, max_computes=0):
-        self.total_computes = 0
-        self.max_computes = max_computes
-
-    def __call__(self, dsk, keys, **kwargs):
-        self.total_computes += 1
-        if self.total_computes > self.max_computes:
-            raise RuntimeError(
-                "Too many computes. Total: %d > max: %d."
-                % (self.total_computes, self.max_computes)
-            )
-        return dask.get(dsk, keys, **kwargs)
-
-
-def raise_if_dask_computes(max_computes=0):
-    """
-    Return a dummy context manager so
-    that this can be used for non-dask objects
-    """
-    scheduler = CountingScheduler(max_computes)
-    return dask.config.set(scheduler=scheduler)
 
 
 def drop_nans(a, b, weights=None, dim="time"):